Fix tiny-solver forcing rayon into multithreading
