+++ factrs-bench/Cargo.toml
@@ -4,20 +4,12 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-tiny-solver = "0.12.0"
 nalgebra = { version = "0.33", features = ["compare"] }
 factrs = { version = "0.1.0", path = ".." }
-cxx = { version = "1.0", features = ["c++17"] }
+tiny-solver = { git = "https://github.com/contagon/tiny-solver-rs", branch = "rayon" }
 
 [dev-dependencies]
 diol = { version = "0.8.3", default-features = false }
-
-[build-dependencies]
-cmake = "0.1.52"
-cxx-build = "1.0"
-
-[features]
-cpp = []
 
 [[bench]]
 name = "g2o-2d"
