+++ factrs-bench/benches/g2o-2d.rs
@@ -1,8 +1,11 @@
 use diol::prelude::{black_box, list, Bench, BenchConfig, Bencher};
+
+const DATA_DIR: &str = "../examples/data/";
+
 // ------------------------- factrs ------------------------- //
 use factrs::{core::GaussNewton, traits::Optimizer, utils::load_g20};
 fn factrs(bencher: Bencher, file: &str) {
-    let (graph, init) = load_g20(file);
+    let (graph, init) = load_g20(&format!("{}{}", DATA_DIR, file));
     bencher.bench(|| {
         let mut opt: GaussNewton = GaussNewton::new(graph.clone());
         let mut results = opt.optimize(init.clone());
@@ -14,7 +17,7 @@ fn factrs(bencher: Bencher, file: &str) {
 use tiny_solver::{gauss_newton_optimizer, optimizer::Optimizer as TSOptimizer};
 
 fn tinysolver(bencher: Bencher, file: &str) {
-    let (graph, init) = factrs_bench::tiny_solver::load_g2o(file);
+    let (graph, init) = factrs_bench::load_tiny_g2o(&format!("{}{}", DATA_DIR, file));
     bencher.bench(|| {
         let gn = gauss_newton_optimizer::GaussNewtonOptimizer::new();
         let mut results = gn.optimize(&graph, &init, None);
@@ -26,7 +29,7 @@ fn main() -> std::io::Result<()> {
     let to_run = list![factrs, tinysolver];
 
     let mut bench = Bench::new(BenchConfig::from_args()?);
-    bench.register_many(to_run, ["../examples/data/M3500.g2o"]);
+    bench.register_many(to_run, ["M3500.g2o"]);
     bench.run()?;
 
     Ok(())
